+++ trunk/View/Grenade.h
@@ -14,7 +14,7 @@ public:
 	~Grenade();
 	bool isAlive();
 	void update();
-	void startCountDown(unsigned seconds);
+	void startCountDown(float seconds);
 };
